myproject/requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
dataclasses==0.6
Django==2.2.3
djangorestframework==3.9.4
djongo==1.2.33
kombu==4.6.3
paho-mqtt==1.4.0
pymongo==3.8.0
pytz==2019.1
sqlparse==0.2.4
vine==1.3.0
